Added Cassette Mapper, removed any connections to AA.

+++ domain/src/main/java/andrewtorski/cassette/domain/entity/Cassette.java
@@ -14,56 +14,80 @@ public class Cassette {
 
     //region Private Fields
 
-    /** Unique Identifier of the Cassete. */
-    private int id;
+    /**
+     * Unique Identifier of the Cassete.
+     */
+    private long id;
 
-    /** Title of this Cassette. */
+    /**
+     * Title of this Cassette.
+     */
     private String title;
 
-    /** Description of this Cassette.
-     *  Not required. */
+    /**
+     * Description of this Cassette.
+     * Not required.
+     */
     private String description;
 
-    /** Date and Time of the creation of this Cassette. */
+    /**
+     * Date and Time of the creation of this Cassette.
+     */
     private Date dateTimeOfCreation;
 
-    /** Total length of all Recordings.
-     *  Express in seconds. */
+    /**
+     * Total length of all Recordings.
+     * Express in seconds.
+     */
     private int lengthInMilliseconds;
 
-    /** Was this Cassette compiled to one File. */
+    /**
+     * Was this Cassette compiled to one File.
+     */
     private boolean isCompiled;
 
-    /** Compiled audio file path.
-     *  This string is null or empty if this.wasCompiled is null. */
+    /**
+     * Compiled audio file path.
+     * This string is null or empty if this.wasCompiled is null.
+     */
     private String compiledFilePath;
 
-    /** Compiled audio file.
-     *  This file is essentially all Recordings merged into one file.
-     *  This reference is null, if this.wasCompiled is null. */
+    /**
+     * Compiled audio file.
+     * This file is essentially all Recordings merged into one file.
+     * This reference is null, if this.wasCompiled is null.
+     */
     private File compiledFile;
 
-    /** Date and Time of when this Cassette was compiled to one file.
-     *  This reference is null, if this.wasCompiled is null. */
+    /**
+     * Date and Time of when this Cassette was compiled to one file.
+     * This reference is null, if this.wasCompiled is null.
+     */
     private Date dateTimeOfCompilation;
 
-    /** Number of the Recordings on this Cassette. */
+    /**
+     * Number of the Recordings on this Cassette.
+     */
     private int numberOfRecordings;
 
-    /** Collection of Recordings on this Cassette. */
+    /**
+     * Collection of Recordings on this Cassette.
+     */
     private List<Recording> recordings;
 
     //endregion PrivateFields
 
     //region Constructors
 
-    public Cassette(){
+    public Cassette() {
         this.id = -1;
     }
 
-    /** Constructor which should be used when Mapping from DAL Cassette Entity to Domain Cassette. */
-    public Cassette(int id, String title, String description,
-                    Date dateTimeOfCreation,  int lengthInMilliseconds, boolean isCompiled,
+    /**
+     * Constructor which should be used when Mapping from DAL Cassette Entity to Domain Cassette.
+     */
+    public Cassette(long id, String title, String description,
+                    Date dateTimeOfCreation, int lengthInMilliseconds, boolean isCompiled,
                     String compiledFilePath, Date dateTimeOfCompilation,
                     int numberOfRecordings, List<Recording> recordings) {
         this.id = id;
@@ -77,15 +101,17 @@ public class Cassette {
         this.numberOfRecordings = numberOfRecordings;
         this.recordings = recordings;
 
-        if(this.isCompiled){
+        if (this.isCompiled) {
             this.compiledFile = new File(this.compiledFilePath);
-        } else{
+        } else {
             this.compiledFile = null;
         }
     }
 
-    /** Constructors which should be used when mapping from Presentation CassetteModel to Domain
-     *  Cassette. */
+    /**
+     * Constructors which should be used when mapping from Presentation CassetteModel to Domain
+     * Cassette.
+     */
     public Cassette(String title, String description) {
         this.title = title;
         this.description = description;
@@ -96,7 +122,7 @@ public class Cassette {
 
     //region Methods
 
-    public int incrementAndReturnNumberOfRecordings(){
+    public int incrementAndReturnNumberOfRecordings() {
         return ++numberOfRecordings;
     }
 
@@ -104,11 +130,11 @@ public class Cassette {
 
     //region Getters & Setters
 
-    public int getId() {
+    public long getId() {
         return id;
     }
 
-    public void setId(int id) {
+    public void setId(long id) {
         this.id = id;
     }
 
@@ -144,7 +170,7 @@ public class Cassette {
         this.lengthInMilliseconds = lengthInMilliseconds;
     }
 
-    public int getLengthInSeconds(){
+    public int getLengthInSeconds() {
         return this.lengthInMilliseconds / 1000;
     }
